코드/requirements.txt
mysql-connector-python==8.0.33
pandas==2.2.1
openpyxl==3.1.2
requests==2.31.0 